% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{caretta}
\alias{caretta}
\title{A data.frame created via the read.table function}
\format{
dataframe
}
\usage{
caretta
}
\description{
This a test data set used for the dartR tutorial A Primer on R for Windows.
}
\author{
Emily but originally provided by Arthur Georges (bugs? Post to
 \url{https://groups.google.com/d/forum/dartr}
}
\keyword{datasets}
