% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{gl_sample}
\alias{gl_sample}
\title{A genlight object created via the read.dart functions}
\format{
genlight object
}
\usage{
gl_sample
}
\description{
This a test data set used for the dartR tutorial on data 
structure and dat a input. It is a genlight object of freshwater turtles.
}
\author{
Emily but originally provided by Arthur Georges (bugs? Post to
 \url{https://groups.google.com/d/forum/dartr}
}
\keyword{datasets}
